format

+++ src/legend-scripts/src/interfaces/IComet.sol
@@ -2,7 +2,13 @@
 pragma solidity 0.8.23;
 
 interface IComet {
-    event AbsorbCollateral(address indexed absorber, address indexed borrower, address indexed asset, uint collateralAbsorbed, uint usdValue);
+    event AbsorbCollateral(
+        address indexed absorber,
+        address indexed borrower,
+        address indexed asset,
+        uint256 collateralAbsorbed,
+        uint256 usdValue
+    );
 
     function getAssetInfo(uint8 i) external view returns (AssetInfo memory);
